requirements.txt
streamlit
openai-whisper
transformers
torch
moviepy
scikit-learn
pandas